fix: improve QwikCity <head>
