--- starters/apps/qwik-city/src/components/head/social.tsx
@@ -1,10 +1,0 @@
-export const Social = () => {
-  return (
-    <>
-      <meta property="og:site_name" content="Qwik" />
-      <meta name="twitter:card" content="summary_large_image" />
-      <meta name="twitter:site" content="@QwikDev" />
-      <meta name="twitter:title" content="Qwik" />
-    </>
-  );
-};
