+++ starters/apps/qwik-city/src/root.tsx
@@ -1,13 +1,22 @@
 import { component$ } from '@builder.io/qwik';
 import { QwikCity, RouterOutlet, ServiceWorkerRegister } from '@builder.io/qwik-city';
-import { Head } from './components/head/head';
+import { RouterHead } from './components/head/head';
 
 import './global.css';
 
 export default component$(() => {
+  /*
+   * The root of a QwikCity site always start with the <QwikCity> component,
+   * immediately followed by the document's <head> and <body>.
+   *
+   * Dont remove the `<head>` and `<body>` elements.
+   */
   return (
     <QwikCity>
-      <Head />
+      <head>
+        <meta charSet="utf-8" />
+        <RouterHead />
+      </head>
       <body lang="en">
         <RouterOutlet />
         <ServiceWorkerRegister />
